Fixing opening relative org board links
Both eww-open-file and find-name-dired depend on the value of default-directory, not always equivalent to the current buffer directory, to expand relative links because they use expand-file-name without passing a directory argument. The filename passed to them has to be expanded as an absolute path in advance so those commands will not depend on the default-directory value. By including the directory of the current buffer in the expand-file-name in the changed lines, it will expand to an absolute folder before being passed to the eww and dired commands.